+++ Tests/Wire/Resources/Shipping/OrderDrafts/Expanded/ExpandedClientWireTests.swift
@@ -23,6 +23,9 @@ import Chrt
                     "forwarder_org_id": "forwarder_org_id",
                     "shipper_org_id": "shipper_org_id",
                     "off_chrt_shipper_org_id": "off_chrt_shipper_org_id",
+                    "courier_org_ids": [
+                      "courier_org_ids"
+                    ],
                     "status": "draft",
                     "draft_started_at_timestamp": "2024-01-15T09:30:00Z",
                     "staged_at_timestamp": "2024-01-15T09:30:00Z",
@@ -153,6 +156,9 @@ import Chrt
                 forwarderOrgId: Optional("forwarder_org_id"),
                 shipperOrgId: Optional("shipper_org_id"),
                 offChrtShipperOrgId: Optional("off_chrt_shipper_org_id"),
+                courierOrgIds: Optional([
+                    "courier_org_ids"
+                ]),
                 status: Optional(.draft),
                 draftStartedAtTimestamp: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
                 stagedAtTimestamp: Optional(try! Date("2024-01-15T09:30:00Z", strategy: .iso8601)),
@@ -405,6 +411,9 @@ import Chrt
             search: "search",
             filterDraftStartedAtTimestampLte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
             filterDraftStartedAtTimestampGte: try! Date("2024-01-15T09:30:00Z", strategy: .iso8601),
+            filterCourierOrgId: "filter_courier_org_id",
+            filterForwarderOrgId: "filter_forwarder_org_id",
+            filterShipperOrgId: "filter_shipper_org_id",
             request: .init(body: OrderAndTaskGroupExpandedReq(
 
             )),